+++ okr-client/src/main.tsx
@@ -4,7 +4,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import App from './App.tsx';
-import OkrProvider from './context/okr.provider.tsx';
+import OkrProvider from './context/okr-data/okr.provider.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
